MXS-2208 Compile avro as C++
An unantecipated sidestep in the path of introducing a small
HTTP library to maxbase.

+++ avro/maxavrocheck.cc
@@ -45,7 +45,7 @@ int check_file(const char* filename)
     if (!dump)
     {
         printf("File sync marker: ");
-        for (int i = 0; i < sizeof(file->sync); i++)
+        for (size_t i = 0; i < sizeof(file->sync); i++)
         {
             printf("%hhx", file->sync[i]);
         }
